Enabled memcached session backend and clear cache when service config changes
The keystone endpoint is cached and may not be valid across config changes.
Toggling keystone TLS in particular will change the keystone url and break
any existing horizon session.

This switches from cookie session backend to memcached backend and uses the
CONFIG_HASH as the cache prefix to ensure caches are invalidated if/when the
config changes.

Jira: OSPRH-4392